Added beginning of volume handling and modified how the solver works.
NEW
---
 * volumehandler.h/cpp: prototype for a new class handling volume evolution through time.

DELETED
-------
 * reactiongroup.cpp: unused implementation file.
 * reactiongroupevent.h: new version moved within `ManualDispatchSolver` implementation.

MAJOR
-----
 * added a `handle_volume_change` method for all reactions.
       * bidirectionalreaction.h, reaction.h: added method to interface.
       * forwardreaction.h, backwardreaction.h: implemented method.
       * loading.h/cpp, sequencebinding.h/cpp, release.h/cpp, translocation.h/cpp: implemented method, added a reaction constant that depends on the volume.

 * changed interface to `Solver` and internal structure. 
   	* solver.h/cpp: (i) An accessor to the next reaction time has beend added. Internally, performing reaction is done by `Solver`, a command to schedule the next reaction and an accessor to next reaction have been added, (ii) A rescheduling method has been added to reset time during simulation, (iii) A method for changing volume has been added.
        * naivesolver.h: implemented new interface.
	* manualdispachsolver.h: (i) implemented new interface, (ii) simplified internal algorithm and created helper methods.
	* reactiongroup.h: changed interface to mimick interface of `Solver`. It is now oriented towards scheduling the next reaction instead of performing it.
	* updatedrategroup.h/cpp: implemented new interface.
	* constantrategroup.h/cpp: (i) implemented new interface, (ii) simplified internal algorithm, which was outdated.

 * simulation.cpp: corrected and simplified simulation creation/loop by creating numerous helper methods. The new loop is event oriented and takes profit of the new scheduling interface of the solver. It simply takes the next event happening between performing an event, a volume change or a reaction and takes necessary actions. Because of the scheduling system, logging and stopping of the simulation occurs at the right time (before the system is actually modified).

 * simulationparams.h/cpp, simulation.cpp: output step is now specified in time, not number of reactions

 * ratemanager.h: (i) moved `compute_all_rates` method to public interface, (ii) changed `random_index` accessor to `random_reaction`.

MINOR
-----
 * rategroup.h: formatting.

BUGFIXES
--------
 * simulation.cpp: see MAJOR. Loop was not correctly implemented, timings of logging, events and reactions were not correctly intertwined.

+++ src/simulationparams.h
@@ -113,9 +113,9 @@ class SimulationParams
 
   /**
    * @brief Accessor to output step.
-   * @return Output step in number of reactions. 10000 if none specified
+   * @return Output step. 1 if none specified
    */
-  int output_step (void) const;
+  double output_step (void) const;
   
   /**
    * @brief Accessor to entities to output.
@@ -203,8 +203,8 @@ private:
   /** @brief Output file name for parameters.. */
   static const std::string _param_file;
 
-  /** @brief Output step in number of reactions. 10000 if none specified. */
-  int _output_step;
+  /** @brief Output step. 1 if none specified. */
+  double _output_step;
   
   /** @brief Initial time tag. */
   static const std::string _output_step_tag;
@@ -364,7 +364,7 @@ inline std::string SimulationParams::reaction_file (void) const
   return _output_dir + "/" + _reaction_file;
 }
 
-inline int SimulationParams::output_step (void) const
+inline double SimulationParams::output_step (void) const
 {
   return _output_step;
 }
